move test_mrs_direct_batch_input to test_tracking.py
